+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 * Added support for `roundTo` field in availability response model
 * Added support for `attributes` field in folder model
+* Added support for icloud as an auth provider
 
 ### 7.1.0 / 2024-02-12
 * Added support for `/v3/connect/tokeninfo` endpoint